refactor foundations
