+++ kotlin/build.gradle
@@ -1,35 +1,16 @@
-plugins {
-    id 'java'
-    id 'org.jetbrains.kotlin.jvm' version '1.3.0'
+buildscript {
+    ext.kotlin_version = '1.3.11'
+    repositories {
+        mavenCentral()
+        jcenter()
+        google()
+    }
+    dependencies {
+        classpath 'com.android.tools.build:gradle:3.3.0'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+    }
 }
 
-apply plugin: 'kotlin-kapt'
-
-group 'FPFundamentals'
-version '1.0-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
-repositories {
-    jcenter()
-    maven { url "https://oss.jfrog.org/artifactory/oss-snapshot-local/" }
-}
-
-dependencies {
-    def arrow_version = "0.9.0-SNAPSHOT"
-
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-
-    implementation "io.arrow-kt:arrow-core-data:$arrow_version"
-    implementation "io.arrow-kt:arrow-effects-extensions:$arrow_version"
-
-    kapt "io.arrow-kt:arrow-annotations-processor:$arrow_version"
-}
-
-compileKotlin {
-    kotlinOptions.jvmTarget = "1.8"
-}
-compileTestKotlin {
-    kotlinOptions.jvmTarget = "1.8"
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
