+++ kotlin/typeclasses/build.gradle
@@ -1,0 +1,31 @@
+apply plugin: 'kotlin'
+apply plugin: 'kotlin-kapt'
+
+group 'com.fortyseven'
+version '1.0-SNAPSHOT'
+
+sourceCompatibility = 1.8
+
+repositories {
+    jcenter()
+    maven { url "https://oss.jfrog.org/artifactory/oss-snapshot-local/" }
+}
+
+dependencies {
+    def arrow_version = "0.9.0-SNAPSHOT"
+
+    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+
+    implementation "io.arrow-kt:arrow-core-data:$arrow_version"
+    implementation "io.arrow-kt:arrow-effects-extensions:$arrow_version"
+
+    kapt "io.arrow-kt:arrow-annotations-processor:$arrow_version"
+}
+
+compileKotlin {
+    kotlinOptions.jvmTarget = "1.8"
+}
+compileTestKotlin {
+    kotlinOptions.jvmTarget = "1.8"
+}
