+++ kotlin/typeclasses/src/main/kotlin/com/fortyseven/fpfundamentals/Typeclasses.kt
@@ -1,0 +1,6 @@
+package com.fortyseven.fpfundamentals
+
+interface Combinator<A> {
+
+  fun combine(x: A, y: A): A
+}
